@@ -1,24 +1,37 @@
 Gem::Specification.new do |s|
   s.platform    = Gem::Platform::RUBY
   s.name        = 'spree_product_assembly'
-  s.version     = '2.0.0.beta'
+  s.version     = '3.3.0'
   s.summary     = 'Adds oportunity to make bundle of products to your Spree store'
-  s.description = 'Adds oportunity to make bundle of products to your Spree store'
-  s.required_ruby_version = '>= 1.9.3'
+  s.description = s.summary
+  s.required_ruby_version = '>= 2.1.0'
 
   s.author            = 'Roman Smirnov'
-  s.email             = 'roman@railsdog.com'
-  s.homepage          = 'https://github.com/spree/spree-product-assembly'
+  s.email             = 'POMAHC@gmail.com'
+  s.homepage          = 'https://github.com/spree-contrib/spree-product-assembly'
 
   s.files         = `git ls-files`.split("\n")
-  s.test_files    = `git ls-files -- {test,spec,features}/*`.split("\n")
+  s.test_files    = `git ls-files -- spec/*`.split("\n")
   s.require_path = 'lib'
   s.requirements << 'none'
 
-  s.add_dependency('spree', '~> 2.1.0')
-  
-  s.add_development_dependency 'rspec-rails', '~> 2.14.0'
-  s.add_development_dependency 'sqlite3'
+  s.add_dependency 'spree_backend', '~> 2.1.0'
+
+
+  s.add_development_dependency 'active_model_serializers'
+  s.add_development_dependency 'capybara'
+  s.add_development_dependency 'capybara-screenshot'
+  s.add_development_dependency 'coffee-rails'
+  s.add_development_dependency 'database_cleaner'
+  s.add_development_dependency 'factory_bot'
   s.add_development_dependency 'ffaker'
-  s.add_development_dependency 'factory_girl_rails', '~> 4.2.1'
+  s.add_development_dependency 'launchy'
+  s.add_development_dependency 'pg', '~> 0.18'
+  s.add_development_dependency 'poltergeist'
+  s.add_development_dependency 'rspec-rails'
+  s.add_development_dependency 'sass-rails'
+  s.add_development_dependency 'simplecov'
+  s.add_development_dependency 'sqlite3'
+  s.add_development_dependency 'mysql2'
+  s.add_development_dependency 'appraisal'
 end
