@@ -1,12 +1,9 @@
 Spree::Product.class_eval do
-  has_and_belongs_to_many  :parts, :class_name => "Spree::Variant",
-        :join_table => "spree_assemblies_parts",
-        :foreign_key => "assembly_id", :association_foreign_key => "part_id"
+  has_many :assemblies_parts, through: :variants_including_master,
+           source: :parts_variants
+  has_many :parts, through: :assemblies_parts
 
-  has_many :assemblies_parts, :class_name => "Spree::AssembliesPart",
-    :foreign_key => "assembly_id"
-
-  scope :individual_saled, -> { where(["spree_products.individual_sale = ?", true]) }
+  scope :individual_saled, -> { where(individual_sale: true) }
 
   scope :search_can_be_part, ->(query){ not_deleted.available.joins(:master)
     .where(arel_table["name"].matches("%#{query}%").or(Spree::Variant.arel_table["sku"].matches("%#{query}%")))
@@ -14,45 +11,10 @@ Spree::Product.class_eval do
     .limit(30)
   }
 
-  scope :active, lambda { |*args|
-    not_deleted.individual_saled.available(nil, args.first)
-  }
+  validate :assembly_cannot_be_part, if: :assembly?
 
-  validate :assembly_cannot_be_part, :if => :assembly?
-
-  def add_part(variant, count = 1)
-    ap = Spree::AssembliesPart.get(self.id, variant.id)
-    if ap
-      ap.count += count
-      ap.save
-    else
-      self.parts << variant
-      set_part_count(variant, count) if count > 1
-    end
-  end
-
-  def remove_part(variant)
-    ap = Spree::AssembliesPart.get(self.id, variant.id)
-    unless ap.nil?
-      ap.count -= 1
-      if ap.count > 0
-        ap.save
-      else
-        ap.destroy
-      end
-    end
-  end
-
-  def set_part_count(variant, count)
-    ap = Spree::AssembliesPart.get(self.id, variant.id)
-    unless ap.nil?
-      if count > 0
-        ap.count = count
-        ap.save
-      else
-        ap.destroy
-      end
-    end
+  def variants_or_master
+    has_variants? ? variants : [master]
   end
 
   def assembly?
@@ -60,11 +22,17 @@ Spree::Product.class_eval do
   end
 
   def count_of(variant)
-    ap = Spree::AssembliesPart.get(self.id, variant.id)
-    ap ? ap.count : 0
+    ap = assemblies_part(variant)
+    # This checks persisted because the default count is 1
+    ap.persisted? ? ap.count : 0
   end
 
   def assembly_cannot_be_part
-    errors.add(:can_be_part, Spree.t(:assembly_cannot_be_part)) if can_be_part
+    errors.add(:can_be_part, Spree.t(:assembly_cannot_be_part)) if can_be_part?
+  end
+
+  private
+  def assemblies_part(variant)
+    Spree::AssembliesPart.get(self.id, variant.id)
   end
 end
