@@ -1,4 +1,4 @@
-class NamespaceProductAssemblyForSpreeOne < ActiveRecord::Migration
+class NamespaceProductAssemblyForSpreeOne < SpreeExtension::Migration[4.2]
   def up
     rename_table :assemblies_parts, :spree_assemblies_parts
   end
