@@ -1,17 +1,15 @@
-require 'spec_helper'
-
 module Spree
-  describe AssembliesPart do
+  describe AssembliesPart, type: :model do
     let(:product) { create(:product) }
     let(:variant) { create(:variant) }
 
     before do
-      product.parts.push variant
+      product.master.parts.push variant
     end
 
     context "get" do
       it "brings part by product and variant id" do
-        subject.class.get(product.id, variant.id).part.should == variant
+        expect(subject.class.get(product.id, variant.id).part).to eq variant
       end
     end
   end
