@@ -3,22 +3,23 @@ module Spree
     # Overridden from spree core to make it also check for assembly parts stock
     class AvailabilityValidator < ActiveModel::Validator
       def validate(line_item)
-        product = line_item.product
+        line_item.quantity_by_variant.each do |variant, variant_quantity|
+          inventory_units = line_item.inventory_units.where(variant: variant).count
+          quantity = variant_quantity - inventory_units
 
-        valid = if product.assembly?
-          line_item.parts.all? do |part|
-            Stock::Quantifier.new(part).can_supply?(line_item.count_of(part) * line_item.quantity)
+          next if quantity <= 0
+
+          quantifier = Stock::Quantifier.new(variant)
+
+          unless quantifier.can_supply? quantity
+            display_name = %Q{#{variant.name}}
+            display_name += %Q{ (#{variant.options_text})} unless variant.options_text.blank?
+
+            line_item.errors[:quantity] << Spree.t(
+              :selected_quantity_not_available,
+              item: display_name.inspect
+            )
           end
-        else
-          Stock::Quantifier.new(line_item.variant).can_supply? line_item.quantity
-        end
-
-        unless valid
-          variant = line_item.variant
-          display_name = %Q{#{variant.name}}
-          display_name += %Q{ (#{variant.options_text})} unless variant.options_text.blank?
-
-          line_item.errors[:quantity] << Spree.t(:out_of_stock, :scope => :order_populator, :item => display_name.inspect)
         end
       end
     end
