@@ -1,5 +1,6 @@
 Deface::Override.new(:virtual_path => "spree/admin/products/_form",
                      :name => "product_assembly_admin_product_form_right",
-                     :insert_after => "[data-hook='admin_product_form_right'], #admin_product_form_right[data-hook]",
+                     :insert_after => "[data-hook='admin_product_form_description']",
                      :partial => "spree/admin/products/product_assembly_fields",
+                     :original => "eb9ecf7015fa51bb0adf7dafd7e6fdf1d652025d",
                      :disabled => false)
