@@ -1,4 +1,4 @@
-Spree::Core::Engine.routes.append do
+Spree::Core::Engine.add_routes do
 
   namespace :admin do
     resources :products do
