@@ -6,8 +6,18 @@ module SpreeProductAssembly
 
     def self.activate
       Dir.glob(File.join(File.dirname(__FILE__), "../../app/**/*_decorator.rb")) do |c|
-        Rails.env.production? ? require(c) : load(c)
+        Rails.configuration.cache_classes ? require(c) : load(c)
       end
+
+      if ::Rails::Engine.subclasses.map(&:name).include? "Spree::Wombat::Engine"
+        Dir.glob(File.join(File.dirname(__FILE__), "../../lib/**/*_serializer.rb")) do |serializer|
+          Rails.env.production? ? require(serializer) : load(serializer)
+        end
+      end
+    end
+
+    initializer 'spree.assets.precompile', group: :all do |app|
+      app.config.assets.precompile += %w( spinner.gif )
     end
 
     config.to_prepare &method(:activate).to_proc
