@@ -1,38 +1,34 @@
-# Configure Rails Environment
-ENV["RAILS_ENV"] = "test"
+require 'simplecov'
+SimpleCov.start 'rails'
 
-require File.expand_path("../dummy/config/environment.rb",  __FILE__)
+ENV['RAILS_ENV'] = 'test'
+
+begin
+  require File.expand_path('../dummy/config/environment', __FILE__)
+rescue LoadError
+  puts 'Could not load dummy application. Please ensure you have run `bundle exec rake test_app`'
+  exit
+end
+
 require 'rspec/rails'
 require 'ffaker'
 
-# Requires supporting ruby files with custom matchers and macros, etc,
-# in spec/support/ and its subdirectories.
-Dir[File.join(File.dirname(__FILE__), "support/**/*.rb")].each {|f| require f }
-
-require 'spree/testing_support/factories'
-require 'spree/testing_support/url_helpers'
-require 'spree/testing_support/authorization_helpers'
-
 RSpec.configure do |config|
-  config.color = true
+  config.fail_fast = false
+  config.filter_run focus: true
+  config.infer_spec_type_from_file_location!
   config.mock_with :rspec
+  config.raise_errors_for_deprecations!
+  config.run_all_when_everything_filtered = true
+  config.use_transactional_fixtures = false
 
-  config.use_transactional_fixtures = true
-
-  config.include FactoryGirl::Syntax::Methods
-
-  config.extend Spree::TestingSupport::AuthorizationHelpers::Request, type: :feature
-end
-
-class ActiveRecord::Base
-  mattr_accessor :shared_connection
-  @@shared_connection = nil
-
-  def self.connection
-    @@shared_connection || retrieve_connection
+  config.expect_with :rspec do |expectations|
+    expectations.syntax = :expect
   end
 end
 
-# Forces all threads to share the same connection. This works on
-# Capybara because it starts the web server in a thread.
-ActiveRecord::Base.shared_connection = ActiveRecord::Base.connection
+Dir[File.join(File.dirname(__FILE__), '/support/**/*.rb')].each do |file|
+  require file unless file.include? 'capybara'
+end
+
+require File.join(File.dirname(__FILE__), '/support/capybara.rb')
